storage: add automatic expiry to MongoDB
This changes the record layout in MongoDB so that the modification time
is now stored as a datetime object instead of a timestamp. It enables us
to create an auto-expiry index on that field.

Note that expiring objects is an asynchronous task in MongoDB. Unlike
Redis, MongoDB uses collection-level expiry settings, therefore it is
not possible to create multiple wrapper objects that operate on the same
collection with different TTL values.

Additionally, there is a minor change that allows specifying the
database name in the URI instead of an argument. In case the name is
specified both in the URI and the argument, the one in the argument is
silently ignored.